Beginning of generated documentation using Sphnix
This commit adds a new folder: doc. This folder contains the
doucmentation of this project. It is generated to various formats thanks
to Sphnix. This commit has some sections ported over from the README and
can even be build using CMake.

However, this is not complete yet. For example, only singlehtml works
and i have plans for better manpages (wla-dx, wlalink.1, wla-CPU.1..)