make: POSIX Makefiles support
This commit adds a Makefile to the repository that can be used to build
WLA DX. It is located in mk/generate_mk.sh, which is a unix shell script
that generates the Makefile from a set of information, like what
architectures with what defines should be used. Any changes to it should
regenerate the Makefile and should be commited.

Stuff missing:

* Building documentation
* Version information
* Testing

make: Made it working, missing version information