Fix that refs from special sections aren't found when discarding
In the z80/linker_header_test, there is a section called BANKHEADER,
that is treated specially and put into its own list.

However, when determining what section is used by the linker (so it can
remove that section), it checks all references with each section.
However, a reference from that special BANKHEADER section won't be found
in the normal section list, where it looked up. And since there was an
assumption that there is at least one reference to a section, there was
no NULL check and it just walks past the list.

Fixed by checking the special list first.