Set `tmp_name` to `NULL` before exiting
While trying out DJGPP-GCC I've noticed that the program crashes after
printing the help. I then noticed in GDB it was crashing when trying to
execute `remove(tmp_name)`, and I've noticed that `tmp_name` is only set
some time after checking whenever trying to print the help or not.

This just sets `tmp_name` to be `NULL` by default.