fix ‘parse_flags_result’ may be used uninitialized
warning: ‘parse_flags_result’ may be used uninitialized in this function [-Wmaybe-uninitialized]

It seems gcc thinks that there might be a way to read
`parse_flags_result` uninitalizd when `output_format` is not
`OUTPUT_NONE`, however the only way `output_format` is not `OUTPUT_NONE`
is when `parse_flags` is setting it, but the result of that function
gets saved to `parse_flags_result` anyway, so it shouldn't be really
possible.

Anyway, just fixing it to make gcc happy.