spc700 1-bit opcode fix
On instructions such as bbc, set1, clr1, if the address operand was not
a constant number (eg $40), the compilation would fail with "Unknown symbol."

I tracked this bug coming from how the parser was not splitting the
address argument from the bit specifier.

Cool note: The bit specifier itself does not need to be a strict constant
(0-7), fortunately we can also use a definition, for example.

    set1 Flags.bPlaySong ; sets the bPlaySong bit in Flags