Fix CMake build as a subdirectory
I am embedding wla-dx into the source tree of my build, and I figured out that CMake wasn't building. This is because some of the locations of generated files are coded to CMAKE_BINARY_DIR, when they should be CMAKE_CURRENT_BINARY_DIR.

This PR should fix the issue. It builds correctly on my machine, hoping Azure works as well.