Add support for multiple include directories via the -I flag
This change allows the command line parsing to support multiple uses of the -I flag, allowing for multiple include directories to be added.

When searching for includes, the old pattern of priorities applies (External, incdir, current), however the externals are now processed in order of inclusion on the command line.

It is backwards compatible with existing uses of the -I flag.